Move unused org scan history to its own file
